Make cholesky(::Gaussian) return a Cholesky object
This ensures consistency with LinearAlgebra, in which calling `cholesky`
on a matrix produces a `Cholesky` object. The upper (or lower)
triangular matrix from the factorization can then be accessed via the
`U` or `L` property.

This requires raising the minimum Nabla version to 0.6.0, as that
version introduced support for explicit `Cholesky` constructors.