Simplify Gaussian constructors
In order to simplify the `Gaussian` constructor, we can introduce an
internal function which computes the appropriate types to store, then
use that in an inner constructor. We can now have just a single
definition for `Gaussian`, with the exception of one used for
deprecation purposes.

Note that this requires bumping the dependency on Nabla to 0.8.0, which
introduces sensitivies for `copy`, which we need to materialize the lazy
`Adjoint` and `Transpose` wrappers.

Included in this commit is also a fix for `show`ing `Gaussian`s, which
is needed because `nothing` cannot be printed, and the default `show`
method for `Distribution` subtypes calls `print` rather than `show` on
the fields.