Remove dependency on Nullables
Currently the Nullables package is only being used in one spot, to
denote that the value for the `name` of a parameter may not be present.
The modern way to represent this is with `nothing` to denote the absence
of a value, so the return type of `name` is now `Union{String,
Nothing}`.