Remove use of at-view
Sensitivities for `view` are not implemented in Nabla. (In fact, they
are partially implemented but commented out with a note.) When
optimizing `H`, this code ends up being differentiated by Nabla, since
`H` determines `P` and `yl = y * P'`. We do not currently optimize `H`,
which is why this was not noticed when initially implemented.

Removing `@view` means we'll allocate a copy, but unfortunately we can't
really do better at the moment.