Minor improvement to sq_pairwise_dist sensitivities
There are a lot of transposes happening, but we can use properties of
the transpose operator to simplify the expressions and reduce the number
of transposes.

Benchmarking shows no difference in allocations from that implemented in
the previous commit, but a very small improvement to performance.
Technical differences aside, this form is subjectively easier to read.