Simplify isapprox(::Parameter, ::Parameter)
Fixes #7.

This uses dispatch as a fast-path to returning `false` when comparing
parameters of different types, and it exploits the fact that two
parameters of the same type must have the same fields. Lastly, it uses
an internal `_isapprox` function which has a method for comparing
strings, necessary for named parameters, that allows us to remove
another case of type piracy (`isapprox(::String, ::String)`).